Refactored Script Editor module
- Separated classes into different files
- Fixed little inconsistencies in the DefaultScriptEditor
- Created 'ScriptStyling' Interface
- Implemented 'Plain' and 'Groovy' stylings

- TODO: Make the console's styling dependent on the ScriptStyling too.
