+++ qupath-extension-script-editor/src/main/java/qupath/lib/gui/scripting/richtextfx/RichScriptEditor.java
@@ -33,13 +33,11 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.util.concurrent.ExecutionException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import org.fxmisc.flowless.VirtualizedScrollPane;
 import org.fxmisc.richtext.CodeArea;
 import org.fxmisc.richtext.LineNumberFactory;
 import org.fxmisc.richtext.model.StyleSpans;
@@ -47,26 +45,20 @@ import org.fxmisc.richtext.model.StyleSpansBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javafx.beans.property.ReadOnlyBooleanProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
-import javafx.beans.value.ObservableValue;
-import javafx.concurrent.Task;
+import javafx.beans.binding.Bindings;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleObjectProperty;
 import javafx.scene.control.ContextMenu;
-import javafx.scene.control.IndexRange;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCodeCombination;
 import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
-import javafx.scene.layout.Region;
-import qupath.lib.common.ThreadTools;
 import qupath.lib.gui.QuPathGUI;
 import qupath.lib.gui.scripting.DefaultScriptEditor;
 import qupath.lib.gui.scripting.QPEx;
+import qupath.lib.gui.scripting.ScriptEditorControl;
 import qupath.lib.gui.tools.MenuTools;
 import qupath.lib.scripting.QP;
-
-
 
 /*
  * 
@@ -88,7 +80,7 @@ import qupath.lib.scripting.QP;
 
 /**
  * 
- * Rich text script editor for QuPath.
+ * Rich script editor for QuPath, which can be used for handling different languages.
  * <p>
  * Makes use of RichTextFX, Copyright (c) 2013-2017, Tomas Mikula and contributors (BSD 2-clause license).
  * 
@@ -99,26 +91,21 @@ public class RichScriptEditor extends DefaultScriptEditor {
 	
 	final private static Logger logger = LoggerFactory.getLogger(RichScriptEditor.class);
 	
-	private static final String[] KEYWORDS = new String[] {
-            "abstract", "assert", "boolean", "break", "byte",
-            "case", "catch", "char", "class", "const",
-            "continue", "default", "do", "double", "else",
-            "enum", "extends", "final", "finally", "float",
-            "for", "goto", "if", "implements", "import",
-            "instanceof", "int", "interface", "long", "native",
-            "new", "null", "package", "private", "protected", "public",
-            "return", "short", "static", "strictfp", "super",
-            "switch", "synchronized", "this", "throw", "throws",
-            "transient", "try", "void", "volatile", "while",
-            "def", "in", "with", "trait", "true", "false", "var" // Groovy
-    };
+	// Store the method names for auto-completion
+	private static final Set<String> METHOD_NAMES = new HashSet<>();
 	
+	private static Pattern PATTERN_CONSOLE;
+	
+	private static KeyCodeCombination completionCode = new KeyCodeCombination(KeyCode.SPACE, KeyCombination.CONTROL_DOWN);
+	
+	final ObjectProperty<ScriptStyling> scriptStyling = new SimpleObjectProperty<>();
+
 	// Delay for async formatting, in milliseconds
 	private static int delayMillis = 100;
+		
+	private AutoCompletor completor;	
+	private ContextMenu menu;
 	
-	private static Pattern PATTERN;
-	private static Pattern PATTERN_CONSOLE;
-	private static final Set<String> METHOD_NAMES = new HashSet<>();
 	static {
 		for (Method method : QPEx.class.getMethods()) {
 			// Exclude deprecated methods (don't want to encourage them...)
@@ -171,44 +158,14 @@ public class RichScriptEditor extends DefaultScriptEditor {
 		METHOD_NAMES.add("print");
 		METHOD_NAMES.add("println");
 		
-		final String KEYWORD_PATTERN = "\\b(" + String.join("|", KEYWORDS) + ")\\b";
-//		final String METHOD_PATTERN = "[a-zA-Z]+\\(";
-//		final String METHOD_PATTERN = "\\b(" + String.join("|", METHOD_NAMES) + ")\\b";
-	    final String PAREN_PATTERN = "\\(|\\)";
-	    final String BRACE_PATTERN = "\\{|\\}";
-	    final String BRACKET_PATTERN = "\\[|\\]";
-	    final String SEMICOLON_PATTERN = "\\;";
-	    final String TRIPLE_QUOTE_PATTERN = "\"\"\"([^\"\"\"\\\\]|\\\\.)*\"\"\"";
-	    final String DOUBLE_QUOTE_PATTERN = "\"([^\"\\\\]|\\\\.)*\"";
-	    final String SINGLE_QUOTE_PATTERN = "'([^'\\\\]|\\\\.)*\'";
-	    final String COMMENT_PATTERN = "//[^\n]*" + "|" + "/\\*(.|\\R)*?\\*/";
-	    
-	    PATTERN = Pattern.compile(
-	            "(?<KEYWORD>" + KEYWORD_PATTERN + ")"
-//	            + "|(?<METHOD>" + METHOD_PATTERN + ")"
-	            + "|(?<PAREN>" + PAREN_PATTERN + ")"
-	            + "|(?<BRACE>" + BRACE_PATTERN + ")"
-	            + "|(?<BRACKET>" + BRACKET_PATTERN + ")"
-	            + "|(?<SEMICOLON>" + SEMICOLON_PATTERN + ")"
-	            + "|(?<TRIPLEQUOTES>" + TRIPLE_QUOTE_PATTERN + ")"
-	            + "|(?<DOUBLEQUOTES>" + DOUBLE_QUOTE_PATTERN + ")"
-	            + "|(?<SINGLEQUOTES>" + SINGLE_QUOTE_PATTERN + ")"
-	            + "|(?<COMMENT>" + COMMENT_PATTERN + ")"
-	    );
-	    
-	    
-	    final String WARNING_PATTERN = "WARN[^\n]*";
+		final String WARNING_PATTERN = "WARN[^\n]*";
 	    final String ERROR_PATTERN = "ERROR:[^\n]*";
 	    
-	    PATTERN_CONSOLE = Pattern.compile(
+		PATTERN_CONSOLE = Pattern.compile(
 	            "(?<ERROR>" + ERROR_PATTERN + ")"
 	            + "|(?<WARN>" + WARNING_PATTERN + ")"
 	    );
 	}
-	
-	private ExecutorService executor = Executors.newSingleThreadExecutor(ThreadTools.createThreadFactory("rich-text-highlighting", true));
-	
-	private ContextMenu menu;
 	
 	/**
 	 * Constructor.
@@ -216,7 +173,6 @@ public class RichScriptEditor extends DefaultScriptEditor {
 	 */
 	public RichScriptEditor(QuPathGUI qupath) {
 		super(qupath);
-		
 		menu = new ContextMenu();
 		MenuTools.addMenuItems(menu.getItems(),
 				MenuTools.createMenu("Run...",
@@ -240,6 +196,7 @@ public class RichScriptEditor extends DefaultScriptEditor {
 	protected ScriptEditorControl getNewEditor() {
 		try {
 			CodeArea codeArea = new CustomCodeArea();
+			CodeAreaControl control = new CodeAreaControl(codeArea);
 			
 			/*
 			 * Using LineNumberFactory.get(codeArea) gives errors related to the new paragraph folding introduced in RichTextFX 0.10.6.
@@ -259,37 +216,6 @@ public class RichScriptEditor extends DefaultScriptEditor {
 			
 			codeArea.setStyle("-fx-background-color: -fx-control-inner-background;");
 			
-			
-			codeArea.setOnContextMenuRequested(e -> {
-				menu.show(codeArea.getScene().getWindow(), e.getScreenX(), e.getScreenY());
-//				menu.show(codeArea, e.getScreenX(), e.getScreenY());
-			});
-
-			
-			CodeAreaControl control = new CodeAreaControl(codeArea);
-			
-			@SuppressWarnings("unused")
-			var cleanup = codeArea
-					.multiPlainChanges()
-					.successionEnds(Duration.ofMillis(delayMillis))
-					.supplyTask(() -> computeHighlightingAsync(codeArea.getText()))
-					.awaitLatest(codeArea.multiPlainChanges())
-					.filterMap(t -> {
-						if (t.isSuccess()) {
-							return Optional.of(t.get());
-						} else {
-							var exception = t.getFailure();
-							String message = exception.getLocalizedMessage() == null ? exception.getClass().getSimpleName() : exception.getLocalizedMessage();
-							logger.error("Error applying syntax highlighting: {}", message);
-							logger.debug("{}", t);
-							return Optional.empty();
-						}
-					})
-					.subscribe(change -> codeArea.setStyleSpans(0, change));
-			
-			
-			codeArea.getStylesheets().add(getClass().getClassLoader().getResource("scripting_styles.css").toExternalForm());
-			
 			// Catch key typed events for special character handling (which should be platform-agnostic)
 			codeArea.addEventFilter(KeyEvent.KEY_TYPED, e -> {
 				if (e.isConsumed())
@@ -307,7 +233,6 @@ public class RichScriptEditor extends DefaultScriptEditor {
 				} else if ("\'".equals(e.getCharacter())) {
 					handleQuotes(control, false);
 					e.consume();
-					
 				}
 				if (!e.isConsumed())
 					matchMethodName(control, e);
@@ -324,7 +249,7 @@ public class RichScriptEditor extends DefaultScriptEditor {
 				} else if (e.isShortcutDown() && e.getCode() == KeyCode.SLASH) {
 					handleLineComment(control);
 					e.consume();
-				} else if (e.getCode() == KeyCode.ENTER && control.getSelectedText().length() == 0) {
+				} else if (e.getCode() == KeyCode.ENTER && codeArea.getSelectedText().length() == 0) {
 					handleNewLine(control);
 					e.consume();
 				} else if (e.getCode() == KeyCode.BACK_SPACE) {
@@ -336,6 +261,53 @@ public class RichScriptEditor extends DefaultScriptEditor {
 					matchMethodName(control, e);
 			});
 
+			codeArea.setOnContextMenuRequested(e -> menu.show(codeArea.getScene().getWindow(), e.getScreenX(), e.getScreenY()));
+			
+			@SuppressWarnings("unused")
+			var cleanup = codeArea
+					.multiPlainChanges()
+					.successionEnds(Duration.ofMillis(delayMillis))
+					.supplyTask(() -> scriptStyling.get().computeHighlightingAsync(codeArea.getText()))
+					.awaitLatest(codeArea.multiPlainChanges())
+					.filterMap(t -> {
+						if (t.isSuccess())
+							return Optional.of(t.get());
+						
+						var exception = t.getFailure();
+						String message = exception.getLocalizedMessage() == null ? exception.getClass().getSimpleName() : exception.getLocalizedMessage();
+						logger.error("Error applying syntax highlighting: {}", message);
+						logger.debug("{}", t);
+						return Optional.empty();
+					})
+					.subscribe(change -> codeArea.setStyleSpans(0, change));
+
+			codeArea.getStylesheets().add(getClass().getClassLoader().getResource("scripting_styles.css").toExternalForm());
+			
+			scriptStyling.bind(Bindings.createObjectBinding(() -> {
+				Language l = getCurrentLanguage();
+				if (l == null)
+					return null;
+				switch(l) {
+				case GROOVY:
+					return new GroovyStyling();
+				case PLAIN:
+				default:
+					return new PlainStyling();
+				}
+			}, currentLanguage));
+			
+			// Triggered whenever the script styling changes (e.g. change of language)
+			scriptStyling.addListener((v, o, n) -> {
+				if (n == null)
+					return;
+				try {
+					StyleSpans<Collection<String>> changes = scriptStyling.get().computeHighlightingAsync(codeArea.getText()).get();
+					codeArea.setStyleSpans(0, changes);
+				} catch (InterruptedException | ExecutionException ex) {
+					logger.warn("Could not apply styling", ex.getLocalizedMessage());
+				}
+			});
+
 			return control;
 		} catch (Exception e) {
 			// Default to superclass implementation
@@ -344,7 +316,32 @@ public class RichScriptEditor extends DefaultScriptEditor {
 		}
 	}
 	
-	private static KeyCodeCombination completionCode = new KeyCodeCombination(KeyCode.SPACE, KeyCombination.CONTROL_DOWN);
+	@Override
+	protected ScriptEditorControl getNewConsole() {
+		try {
+			CodeArea codeArea = new CodeArea();
+			codeArea.setStyle("-fx-background-color: -fx-control-inner-background;");
+			
+			codeArea.richChanges()
+			.successionEnds(Duration.ofMillis(delayMillis))
+			.subscribe(change -> {
+				// If anything was removed, do full reformatting
+				if (change.getRemoved().length() != 0 || change.getPosition() == 0) {
+					if (!change.getRemoved().getText().equals(change.getInserted().getText()))
+						codeArea.setStyleSpans(0, computeConsoleHighlighting(codeArea.getText()));					
+				} else {
+					// Otherwise format only from changed position onwards
+					codeArea.setStyleSpans(change.getPosition(), computeConsoleHighlighting(change.getInserted().getText()));
+				}
+			});
+			codeArea.getStylesheets().add(getClass().getClassLoader().getResource("scripting_styles.css").toExternalForm());
+			return new CodeAreaControl(codeArea);
+		} catch (Exception e) {
+			// Default to superclass implementation
+			logger.error("Unable to create console area", e);
+			return super.getNewEditor();
+		}
+	}
 	
 	/**
 	 * Try to match and auto-complete a method name.
@@ -365,7 +362,23 @@ public class RichScriptEditor extends DefaultScriptEditor {
 	}
 	
 	
-	private AutoCompletor completor;
+	private static StyleSpans<Collection<String>> computeConsoleHighlighting(final String text) {
+        Matcher matcher = PATTERN_CONSOLE.matcher(text);
+        int lastKwEnd = 0;
+        StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
+        while (matcher.find()) {
+            String styleClass =
+                    matcher.group("ERROR") != null ? "error" :
+                    matcher.group("WARN") != null ? "warning" :
+                    null; /* never happens */
+            assert styleClass != null;
+            spansBuilder.add(Collections.emptyList(), matcher.start() - lastKwEnd);
+            spansBuilder.add(Collections.singleton(styleClass), matcher.end() - matcher.start());
+            lastKwEnd = matcher.end();
+        }
+        spansBuilder.add(Collections.emptyList(), text.length() - lastKwEnd);
+        return spansBuilder.create();
+    }
 	
 	/**
 	 * Helper class for toggling through completions.
@@ -423,271 +436,6 @@ public class RichScriptEditor extends DefaultScriptEditor {
 	}
 	
 	
-	
-	@Override
-	protected ScriptEditorControl getNewConsole() {
-		try {
-			CodeArea codeArea = new CodeArea();
-			codeArea.setStyle("-fx-background-color: -fx-control-inner-background;");
-			
-//			var cleanup = codeArea
-//					.multiPlainChanges()
-//					.successionEnds(Duration.ofMillis(500))
-//					.supplyTask(() -> computeConsoleHighlightingAsync(codeArea.getText()))
-//					.awaitLatest(codeArea.multiPlainChanges())
-//					.filterMap(t -> {
-//						if (t.isSuccess()) {
-//							return Optional.of(t.get());
-//						} else {
-//							var exception = t.getFailure();
-//							String message = exception.getLocalizedMessage() == null ? exception.getClass().getSimpleName() : exception.getLocalizedMessage();
-//							logger.error("Error applying syntax highlighting: {}", message);
-//							logger.debug("{}", t);
-//							return Optional.empty();
-//						}
-//					})
-//					.subscribe(change -> codeArea.setStyleSpans(0, change));
-			
-			codeArea.richChanges()
-				.successionEnds(Duration.ofMillis(delayMillis))
-				.subscribe(change -> {
-					// If anything was removed, do full reformatting
-					if (change.getRemoved().length() != 0 || change.getPosition() == 0) {
-						if (!change.getRemoved().getText().equals(change.getInserted().getText()))
-							codeArea.setStyleSpans(0, computeConsoleHighlighting(codeArea.getText()));					
-					} else {
-						// Otherwise format only from changed position onwards
-						codeArea.setStyleSpans(change.getPosition(), computeConsoleHighlighting(change.getInserted().getText()));
-					}
-		        });
-			codeArea.getStylesheets().add(getClass().getClassLoader().getResource("scripting_styles.css").toExternalForm());
-			return new CodeAreaControl(codeArea);
-		} catch (Exception e) {
-			// Default to superclass implementation
-			logger.error("Unable to create console area", e);
-			return super.getNewEditor();
-		}
-	}
-	
-	private Task<StyleSpans<Collection<String>>> computeHighlightingAsync(final String text) {
-		var task = new Task<StyleSpans<Collection<String>>>() {
-			@Override
-			protected StyleSpans<Collection<String>> call() {
-				return computeHighlighting(text);
-			}
-		};
-		executor.execute(task);
-		return task;
-	}
-	
-//	private Task<StyleSpans<Collection<String>>> computeConsoleHighlightingAsync(final String text) {
-//		var task = new Task<StyleSpans<Collection<String>>>() {
-//			@Override
-//			protected StyleSpans<Collection<String>> call() {
-//				return computeConsoleHighlighting(text);
-//			}
-//		};
-//		executor.execute(task);
-//		return task;
-//	}
-	
-	private static StyleSpans<Collection<String>> computeHighlighting(final String text) {
-        Matcher matcher = PATTERN.matcher(text);
-        int lastKwEnd = 0;
-        StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
-        while (matcher.find()) {
-            String styleClass =
-                    matcher.group("KEYWORD") != null ? "keyword" :
-//                    matcher.group("METHOD") != null ? "method" :
-                    matcher.group("PAREN") != null ? "paren" :
-                    matcher.group("BRACE") != null ? "brace" :
-                    matcher.group("BRACKET") != null ? "bracket" :
-                    matcher.group("SEMICOLON") != null ? "semicolon" :
-                    matcher.group("TRIPLEQUOTES") != null ? "string" :
-                    matcher.group("DOUBLEQUOTES") != null ? "string" :
-                    matcher.group("SINGLEQUOTES") != null ? "string" :
-                    matcher.group("COMMENT") != null ? "comment" :
-                    null; /* never happens */
-            assert styleClass != null;
-            spansBuilder.add(Collections.emptyList(), matcher.start() - lastKwEnd);
-            spansBuilder.add(Collections.singleton(styleClass), matcher.end() - matcher.start());
-            lastKwEnd = matcher.end();
-        }
-        spansBuilder.add(Collections.emptyList(), text.length() - lastKwEnd);
-        return spansBuilder.create();
-    }
-	
-	
-	private static StyleSpans<Collection<String>> computeConsoleHighlighting(final String text) {
-        Matcher matcher = PATTERN_CONSOLE.matcher(text);
-        int lastKwEnd = 0;
-        StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
-        while (matcher.find()) {
-            String styleClass =
-                    matcher.group("ERROR") != null ? "error" :
-                    matcher.group("WARN") != null ? "warning" :
-                    null; /* never happens */
-            assert styleClass != null;
-            spansBuilder.add(Collections.emptyList(), matcher.start() - lastKwEnd);
-            spansBuilder.add(Collections.singleton(styleClass), matcher.end() - matcher.start());
-            lastKwEnd = matcher.end();
-        }
-        spansBuilder.add(Collections.emptyList(), text.length() - lastKwEnd);
-        return spansBuilder.create();
-    }
-	
-	
-	
-	
-	static class CodeAreaControl implements ScriptEditorControl {
-		
-		private VirtualizedScrollPane<CodeArea> scrollpane;
-		private CodeArea textArea;
-		private StringProperty textProperty = new SimpleStringProperty();
-		
-		CodeAreaControl(final CodeArea codeArea) {
-			this.textArea = codeArea;
-			textArea.textProperty().addListener((o, v, n) -> textProperty.set(n));
-			textProperty.addListener((o, v, n) -> {
-				if (n.equals(textArea.getText()))
-					return;
-				textArea.clear();
-				textArea.insertText(0, n);
-			});
-			scrollpane = new VirtualizedScrollPane<>(textArea);
-		}
-		
-		@Override
-		public StringProperty textProperty() {
-//			return textArea.textProperty();
-			return textProperty;
-		}
-
-		@Override
-		public void setText(String text) {
-			textArea.clear();
-			textArea.insertText(0, text);
-		}
-
-		@Override
-		public String getText() {
-			return textArea.getText();
-		}
-
-		@Override
-		public ObservableValue<String> selectedTextProperty() {
-			return textArea.selectedTextProperty();
-		}
-
-		@Override
-		public String getSelectedText() {
-			return textArea.getSelectedText();
-		}
-
-		@Override
-		public Region getControl() {
-			return scrollpane;
-		}
-
-		@Override
-		public boolean isUndoable() {
-			return textArea.isUndoAvailable();
-		}
-
-		@Override
-		public boolean isRedoable() {
-			return textArea.isRedoAvailable();
-		}
-
-		@Override
-		public void undo() {
-			textArea.undo();
-		}
-
-		@Override
-		public void redo() {
-			textArea.redo();
-		}
-
-		@Override
-		public void copy() {
-			textArea.copy();
-		}
-
-		@Override
-		public void cut() {
-			textArea.cut();
-		}
-
-		@Override
-		public void paste(String text) {
-			if (text != null)
-				textArea.replaceSelection(text);
-//			textArea.paste();
-		}
-		
-		@Override
-		public void appendText(final String text) {
-			textArea.appendText(text);
-		}
-
-		@Override
-		public void clear() {
-			textArea.clear();
-		}
-		
-		@Override
-		public int getCaretPosition() {
-			return textArea.getCaretPosition();
-		}
-		
-		@Override
-		public void insertText(int pos, String text) {
-			textArea.insertText(pos, text);
-		}
-		
-		@Override
-		public void deleteText(int startIdx, int endIdx) {
-			textArea.deleteText(startIdx, endIdx);
-		}
-		
-		@Override
-		public ReadOnlyBooleanProperty focusedProperty() {
-			return textArea.focusedProperty();
-		}
-
-		@Override
-		public void deselect() {
-			textArea.deselect();
-		}
-
-		@Override
-		public IndexRange getSelection() {
-			return textArea.getSelection();
-		}
-
-		@Override
-		public void selectRange(int startIdx, int endIdx) {
-			textArea.selectRange(startIdx, endIdx);
-		}
-
-		@Override
-		public void setPopup(ContextMenu menu) {
-			textArea.setContextMenu(menu);
-		}
-    
-		@Override
-		public void positionCaret(int index) {
-			textArea.moveTo(index);
-    }
-    
-		@Override
-		public void requestFollowCaret() {
-			textArea.requestFollowCaret();
-		}
-	}
-	
-	
 	static class CustomCodeArea extends CodeArea {
 		
 		/**
@@ -699,8 +447,5 @@ public class RichScriptEditor extends DefaultScriptEditor {
 			if (text != null)
 				replaceSelection(text);
 		}
-		
 	}
-	
-	
 }
